# backend/src/main/java/com/myfitnessapp/model/ejercicio/GrupoMuscular.java
package com.myfitnessapp.model.ejercicio;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Entity
@Table
public class GrupoMuscular {
  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  private Integer id;
  private String nombre;
  public GrupoMuscular(String nombre){
    this.nombre = nombre;
  }
}